Create README.md
to succeed, it must be a users product. What it means is the website is user oriented. Me, the owner, listen to users' feedback 24x7 to tweek website design, forever cycle of feedback-change, feedback-change, to improve, similar to iterate in programming.
commercial end: it has to be open sourced. how to make money? before the website get popular, take commission of bounties. once it get popular, do ad.